more optimizations

+++ avr/cores/megacommand/Adafruit-GFX-Library/Adafruit_GFX.h
@@ -16,7 +16,7 @@ class Adafruit_GFX : public Print {
   Adafruit_GFX(int16_t w, int16_t h); // Constructor
 
   // This MUST be defined by the subclass:
-  virtual void drawPixel(int16_t x, int16_t y, uint16_t color) = 0;
+  virtual void drawPixel(uint8_t x, uint8_t y, uint8_t color) = 0;
 
   // TRANSACTION API / CORE DRAW API
   // These MAY be overridden by the subclass to provide device-specific
